added check to make sure all options have something selected

+++ run-native.sh
@@ -1,6 +1,19 @@
 #!/bin/bash
-python3 -m venv venv
-. venv/bin/activate
-pip install -r src/requirements.txt
-python src/main.py
-rm -r venv
+
+if [[ "$@" == *"--clean"* ]]; then
+    rm -r venv
+fi
+
+if [ ! -d "venv" ]; then
+    python3 -m venv venv
+    . venv/bin/activate
+    pip install -r src/requirements.txt
+else
+    . venv/bin/activate
+fi
+
+if [[ "$@" == *"--debug"* ]]; then
+    python3 src/main.py --debug
+else
+    python src/main.py
+fi
